+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 -   Support `no_verbose` and `regex` in the verbose rules. (#4 and #5)
 -   Support `explicitOnly` in the verbose rules. (#6)
 -   Restructure the verbose rules.
+-   Handle package rename.
 
 ## 0.3.2